+++ relay-server/src/actors/events.rs
@@ -36,7 +36,9 @@ use crate::envelope::{self, AttachmentType, ContentType, Envelope, Item, ItemTyp
 use crate::http::{HttpError, RequestBuilder};
 use crate::metrics::{RelayCounters, RelayHistograms, RelaySets, RelayTimers};
 use crate::service::ServerError;
-use crate::utils::{self, ChunkedFormDataAggregator, FormDataIter, FutureExt, SamplingResult};
+use crate::utils::{
+    self, ChunkedFormDataAggregator, EnvelopeSummary, FormDataIter, FutureExt, SamplingResult,
+};
 
 #[cfg(feature = "processing")]
 use {
@@ -1478,19 +1480,19 @@ impl Handler<HandleEnvelope> for EventManager {
             start_time,
             sampling_project,
         } = message;
-        let event_category = envelope.get_event_category();
 
         let event_id = envelope.event_id();
         let remote_addr = envelope.meta().client_addr();
 
         let scoping = Rc::new(RefCell::new(envelope.meta().get_partial_scoping()));
         let is_received = Rc::new(AtomicBool::from(false));
+        let envelope_summary = Rc::new(RefCell::new(EnvelopeSummary::compute(&envelope)));
 
         let future = project
             .send(CheckEnvelope::fetched(envelope))
             .map_err(ProcessingError::ScheduleFailed)
             .and_then(|result| result.map_err(ProcessingError::ProjectFailed))
-            .and_then(clone!(scoping, |response| {
+            .and_then(clone!(scoping, envelope_summary, |response| {
                 // Use the project id from the loaded project state to account for redirects.
                 let project_id = response.scoping.project_id.value();
                 metric!(set(RelaySets::UniqueProjects) = project_id as i64);
@@ -1499,7 +1501,10 @@ impl Handler<HandleEnvelope> for EventManager {
 
                 let checked = response.result.map_err(ProcessingError::EventRejected)?;
                 match checked.envelope {
-                    Some(envelope) => Ok(envelope),
+                    Some(envelope) => {
+                        envelope_summary.replace(EnvelopeSummary::compute(&envelope));
+                        Ok(envelope)
+                    }
                     None => Err(ProcessingError::RateLimited(checked.rate_limits)),
                 }
             }))
@@ -1533,6 +1538,7 @@ impl Handler<HandleEnvelope> for EventManager {
                     })
                     .map_err(ProcessingError::ScheduleFailed)
                     .flatten()
+                // TODO: Update envelope_summary once the rate-limiting code emits its own outcomes.
             })
             .and_then(clone!(project, |processed| {
                 let rate_limits = processed.rate_limits;
@@ -1670,13 +1676,6 @@ impl Handler<HandleEnvelope> for EventManager {
                     }
                 }
 
-                // Envelopes not containing events (such as standalone attachment uploads or user
-                // reports) should never create outcomes.
-                let category = match event_category {
-                    Some(event_category) => event_category,
-                    None => return,
-                };
-
                 let outcome = error.to_outcome();
                 if let Some(Outcome::Invalid(DiscardReason::Internal)) = outcome {
                     // Errors are only logged for what we consider an internal discard reason. These
@@ -1694,15 +1693,31 @@ impl Handler<HandleEnvelope> for EventManager {
                     return;
                 }
 
+                let envelope_summary = envelope_summary.borrow();
                 if let Some(outcome) = outcome {
-                    outcome_producer.do_send(TrackOutcome {
-                        timestamp: Instant::now(),
-                        scoping: *scoping.borrow(),
-                        outcome,
-                        event_id,
-                        remote_addr,
-                        category,
-                    })
+                    if let Some(category) = envelope_summary.event_category {
+                        outcome_producer.do_send(TrackOutcome {
+                            timestamp: Instant::now(),
+                            scoping: *scoping.borrow(),
+                            outcome: outcome.clone(),
+                            event_id,
+                            remote_addr,
+                            category,
+                            quantity: 1,
+                        });
+                    }
+
+                    if envelope_summary.attachment_quantity > 0 {
+                        outcome_producer.do_send(TrackOutcome {
+                            timestamp: start_time,
+                            scoping: *scoping.borrow(),
+                            outcome,
+                            event_id,
+                            remote_addr,
+                            category: DataCategory::Attachment,
+                            quantity: envelope_summary.attachment_quantity,
+                        });
+                    }
                 }
             })
             .then(move |x, slf, _| {